+++ storybook/cypress/e2e/list-button.cy.ts
@@ -26,10 +26,26 @@ describe("ListButton", () => {
       .get("@dsoListButtonShadow")
       .find(".dso-sublabel")
       .should("contain.text", "Sublabel")
-      .get("@dsoListButtonShadow");
+      .get("@dsoListButtonShadow")
+      .find("#dso-list-button-checkbox")
+      .should("have.attr", "aria-describedby", "sublabel")
+      .get("@dsoListButton")
+      .then(($listButton) => {
+        $listButton.append('<span slot="subcontent">Subcontent</span>');
+      })
+      .get("@dsoListButtonShadow")
+      .find("#dso-list-button-checkbox")
+      .should("have.attr", "aria-describedby", "sublabel description")
+      .get("@dsoListButtonShadow")
+      .find("#description")
+      .should("have.class", "sr-only")
+      .and("contain", "Subcontent")
+      .get("@dsoListButton")
+      .find('> [slot="subcontent"]')
+      .should("have.attr", "aria-hidden", "true");
   });
 
-  it("should render subcontent in slot", () => {
+  it("should render subcontent in slot without prefix", () => {
     cy.get("dso-list-button").invoke("append", `<span slot="subcontent">Subcontent met <strong>HTML</strong></span>`);
 
     cy.get("dso-list-button")
@@ -41,6 +57,28 @@ describe("ListButton", () => {
         cy.get('dso-list-button > [slot="subcontent"]').invoke("get", 0).should("equal", assignedNodes[0]);
       })
       .percySnapshot();
+
+    cy.get("dso-list-button")
+      .shadow()
+      .as("dsoListButtonShadow")
+      .find("#dso-list-button-checkbox")
+      .should("have.attr", "aria-describedby", "description")
+      .get("@dsoListButtonShadow")
+      .find("#description")
+      .should("have.class", "sr-only")
+      .and("contain.html", "<div>Subcontent met <strong>HTML</strong></div>");
+  });
+
+  it("should render subcontent in slot with prefix", () => {
+    cy.get("dso-list-button")
+      .invoke("append", `<span slot="subcontent">Subcontent met <strong>HTML</strong></span>`)
+      .get("dso-list-button")
+      .invoke("prop", "subcontentPrefix", "subcontentPrefix")
+      .shadow()
+      .find("#description")
+      .should("have.class", "sr-only")
+      .and("contain", "subcontentPrefix:")
+      .and("contain.html", "<div>Subcontent met <strong>HTML</strong></div>");
   });
 
   it("should be accessible", () => {