+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ## Next
 
 ### Fixed
-* DropdownMenu: focus trap hamburger menu ([#2392](https://github.com/dso-toolkit/dso-toolkit/issues/2392))
+* Dropdown Menu: focus trap hamburger menu ([#2392](https://github.com/dso-toolkit/dso-toolkit/issues/2392))
 
 ## 🚙 62.9.0
 
@@ -14,7 +14,6 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 * Expandable: Overflow stays hidden when fully expanded ([#2471](https://github.com/dso-toolkit/dso-toolkit/issues/2471))
 * Logo: Fix contrast in ribbon and enable accessibility test ([#2384](https://github.com/dso-toolkit/dso-toolkit/issues/2384))
 * Logo: Responsiveness stuk ([#2489](https://github.com/dso-toolkit/dso-toolkit/issues/2489))
-* DropdownMenu: focus trap hamburger menu ([#2392](https://github.com/dso-toolkit/dso-toolkit/issues/2392))
 
 ### Docs
 * Web Components: Extra interfaces meedocumenteren in readme.md ([#2381](https://github.com/dso-toolkit/dso-toolkit/issues/2381))