+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## Next
 
+### Fixed
+* List Button: Trefwoorden worden niet voorgelezen als onderdeel van het resultaat/werkzaamheid ([#2294](https://github.com/dso-toolkit/dso-toolkit/issues/2294))
+
 ## 🚙 62.9.0
 
 ### Fixed